+++ test/cypress/cypress/integration/step-definitions/management/rules/The-user-clicks-on-the-page-when.js
@@ -5,14 +5,14 @@ const paginatorSelector = getSelector('paginatorSelector', pageName);
 const paginatorSecondSelector = getSelector('paginatorSecondSelector', pageName);
 const paginatorNextSelector = getSelector('paginatorNextSelector', pageName);
 const paginatorPreviousSelector = getSelector('paginatorPreviousSelector', pageName);
-const rulestableSelector = getSelector('rulestableSelector', pageName);
+const paginatorFirstSelector = getSelector('paginatorFirstSelector', pageName);
 
 When('The user clicks on the {} page button', (page) => {
     let selector;
     elementIsVisible(paginatorSelector);
     switch (page) {
         case 'first':
-            selector = paginatorSelector;
+            selector = paginatorFirstSelector;
             break;
         case 'second':
             selector = paginatorSecondSelector;
@@ -26,22 +26,16 @@ When('The user clicks on the {} page button', (page) => {
         default:
             throw new Error(`The page ${page} is not defined`);
     }
-    elementIsVisible(selector);
-    clickElement(selector)
+    // elementIsVisible(selector);
+    clickElement(selector);
     cy.wait(1500);
 });
 
-When('A new set of rules it\'s displayed', () => {
-    cy.wait(1500);
-    elementIsVisible(paginatorSelector);
-});
 
-Then('The first page of rules it displayed', () => {
-    cy.wait(1500);
-    cy.get(rulestableSelector).then(($rules) => {
-        cy.get('@listRulesText').then(($e) => {
-            expect($rules).to.be.equals($e);
-        })
-    })
 
-});
+
+
+
+
+
+
